requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comtypes==1.4.12
contourpy==1.3.2
curl_cffi==0.13.0
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
faiss-cpu==1.12.0
fastapi==0.121.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.59.0
frozendict==2.4.7
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
graphviz==0.21
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multitasking==0.0.12
namex==0.1.0
narwhals==2.13.0
networkx==3.5
numpy==2.0.0
ollama==0.5.3
openai==1.102.0
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
packaging==25.0
pandas==2.2.2
peewee==3.18.3
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.1
pvporcupine==3.0.5
pvrecorder==1.2.7
pyarrow==22.0.0
PyAudio==0.2.14
pycparser==2.23
pydantic==2.11.7
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyttsx3==2.99
pytz==2024.1
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
selection==0.0.21
sentence-transformers==5.1.2
sentry-sdk==2.45.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sklearn-model==0.0.6
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
starlette==0.49.3
streamlit==1.52.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
yfinance==0.2.66
